application/backend/requirements.txt
python-dotenv
flask-cors
SQLAlchemy
Flask-SQLAlchemy
psycopg2-binary
Flask
PyMySQL
nltk
fuzzywuzzy
python-Levenshtein
PyGithub
google-generativeai
Pillow
inflect
